requirements.txt
blinker
click
colorama
Flask
Flask-MySQLdb
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
mysqlclient
Werkzeug
zipp
PyJWT
dicttoxml
pytest